--- src/modules/transaction/components/TransactionSummary/index.js
@@ -1,1 +1,0 @@
-export { default } from './TransactionSummary';
