+++ src/modules/transaction/components/TxSummarizer/footer.js
