+++ src/modules/transaction/manager/transactionSummary.js
@@ -4,7 +4,7 @@ import { compose } from 'redux';
 import { withTranslation } from 'react-i18next';
 import { secondPassphraseStored } from 'src/modules/auth/store/action';
 import { selectActiveToken, selectActiveTokenAccount } from '@common/store';
-import TransactionSummary from '../components/TransactionSummary/TransactionSummary';
+import TxSummarizer from '../components/TxSummarizer';
 
 const mapStateToProps = state => ({
   token: selectActiveToken(state),
@@ -19,4 +19,4 @@ const mapDispatchToProps = {
 export default compose(
   connect(mapStateToProps, mapDispatchToProps),
   withTranslation(),
-)(TransactionSummary);
+)(TxSummarizer);
